Add files via upload
Work in progress- I was watching old TV shows and had the idea that it should be easier to find out how old an actor was when something was made. This app is meant to make it super quick and easy to find that information, rather than having to manually search for a person's date of birth, the release date of a show or movie, and then doing the math.

This project uses the free to access API for TMDB, The Movie DataBase. 

I'm a beginner at development, so I've probably screwed a lot of things up, but I am working on things like this in an effort to learn app development. I plan to develop this app to be more usable and practical and am considering making it more of a multi-purpose navigator for information stored in the API beyond its original intent.